% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{enforceM}
\alias{enforceM}
\title{Force a user-provided matrix to be of type sparseMatrix}
\usage{
enforceM(x)
}
\arguments{
\item{x}{User matrix}
}
\value{
sparseMatrix
}
\description{
Force a user-provided matrix to be of type sparseMatrix
}
\examples{
enforceM(matrix(runif(6),2,3))

}
