% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etopo30}
\alias{etopo30}
\title{ETOPO30 elevation data}
\format{A data frame with 360 rows and 720 columns; etopo[1,1] corresponds to 89.75N, -179.75E,
        etopo30[1,2] to 89.75N, -179.25E, and so on.}
\source{
\url{https://data.europa.eu/euodp/data/dataset/data_world-digital-elevation-model-etopo5}
}
\usage{
etopo30
}
\description{
A dataset containing global elevation on a 0.5 degree grid, based on ETOPO5 (see getETOPO30).
}
\keyword{datasets}
