% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{getCoarseData}
\alias{getCoarseData}
\title{Build a coarser dataset}
\usage{
getCoarseData(data, lat, lon, verbose)
}
\arguments{
\item{data}{Original dataset}

\item{lat}{Vector of latitudes}

\item{lon}{Vector of longitudes}

\item{verbose}{Print statuts to console?}
}
\value{
List with coarser data, lat and lon
}
\description{
Build a coarser dataset
}
