% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo30.R
\name{getCoarseETOPO}
\alias{getCoarseETOPO}
\title{Obtain a coarser version of ETOPO30}
\usage{
getCoarseETOPO(coarsening_factor)
}
\arguments{
\item{coarsening_factor}{Integer >= 1; resolution = 0.5 * coarsening_factor}
}
\value{
Coarser version of ETOPO30
}
\description{
Obtain a coarser version of ETOPO30
}
\examples{
getCoarseETOPO(4)

}
