% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{getColMax}
\alias{getColMax}
\title{Compute max and which.max for columns of a sparse matrix}
\usage{
getColMax(x, cols)
}
\arguments{
\item{x}{sparse matrix}

\item{cols}{column indices}
}
\value{
dense matrix with two columns: max and which.max
}
\description{
Compute max and which.max for columns of a sparse matrix
}
\examples{
getColMax(sparseMatrix(i=1,j=2,x=9), 1:2)

}
