% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{getColumn}
\alias{getColumn}
\title{Fetch one column of a sparse matrix}
\usage{
getColumn(x, col_id)
}
\arguments{
\item{x}{sparse matrix}

\item{col_id}{column index}
}
\value{
List with non-empty rows and non-zero values
}
\description{
Fetch one column of a sparse matrix
}
\examples{
getColumn(sparseMatrix(i=1,j=2,x=9), 2)

}
