% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{getConnectionMatrix}
\alias{getConnectionMatrix}
\title{Build a connection matrix based on a minimum spanning tree}
\usage{
getConnectionMatrix(dual, verbose)
}
\arguments{
\item{dual}{Dual object}

\item{mstree}{Minimum Spanning Tree}

\item{Verbose}{Print status to console}
}
\value{
(4 x nr x nc) x (4 x nr x nc) sparse matrix of connections in dual tree
}
\description{
Build a connection matrix based on a minimum spanning tree
}
