% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{getDual}
\alias{getDual}
\title{Compute the dual graph object}
\usage{
getDual(data, lat, lon, verbose)
}
\arguments{
\item{dataset}{}
}
\value{
List describing the dual graph object
}
\description{
Compute the dual graph object
}
