% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo30.R
\name{getETOPO30}
\alias{getETOPO30}
\title{Compute ETOPO30 from ETOPO5. Download and extract zipped geotiff from
https://data.europa.eu/euodp/data/dataset/data_world-digital-elevation-model-etopo5}
\usage{
getETOPO30(local_tiff)
}
\arguments{
\item{local_tiff}{Path to local ETOPO5 geotiff file ('alwdgg.tiff')}
}
\value{
ETOPO30
}
\description{
Compute ETOPO30 from ETOPO5. Download and extract zipped geotiff from
https://data.europa.eu/euodp/data/dataset/data_world-digital-elevation-model-etopo5
}
