% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{Break path into groups, according to mean and variance}
\usage{
getGroups(data, path, lat, lon)
}
\arguments{
\item{data}{Dataset}

\item{path}{Path}

\item{lat}{Path latitudes}

\item{lon}{Path longitudes}
}
\value{
data.frame with 2 columns: group and mean value of group
}
\description{
Break path into groups, according to mean and variance
}
