% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Compute valid nearest neighbors of each valid point on a grid}
\usage{
getNeighbors(mat, mask_fun = function(x) !is.na(x), verbose)
}
\arguments{
\item{mat}{Matrix of gridpoints}

\item{mask_fun}{Function that identifies valid observations
For sparseMatrix where 0 denotes NA, use
mask_fun = function(x) as.logical(x != 0)}
}
\value{
List with IDs of valid points (old and new) and list (ngb) of their neighbors
}
\description{
Compute valid nearest neighbors of each valid point on a grid
}
\examples{
mat = matrix(runif(6), 2, 3)
mat[2,2] = NA
getNeighbors(mat, verbose=F)
mat[2,2] = 0
getNeighbors(enforceM(mat), function(x) as.logical(x != 0), verbose=F)

}
