% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{getPath}
\alias{getPath}
\title{Compute a path that connects all the points in a matrix}
\usage{
getPath(connection_matrix, verbose)
}
\arguments{
\item{connection_matrix}{Connection matrix for dual graph}
}
\value{
Path that connects all elements in connection matrix
}
\description{
Compute a path that connects all the points in a matrix
}
