% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{getSimilarityMatrix}
\alias{getSimilarityMatrix}
\title{Compute a square similarity matrix based on a raw rectangular matrix
with gridded information and possibly NAs}
\usage{
getSimilarityMatrix(data, neighbors, verbose)
}
\arguments{
\item{data}{NxM matrix with gridded data}

\item{neighbors}{neighborhood structure of dual graph}
}
\value{
(NxM) x (NxM) sparse similarity matrix
}
\description{
Compute a square similarity matrix based on a raw rectangular matrix
with gridded information and possibly NAs
}
\examples{
n = 6
elevation = matrix(1, n, n)
elevation[1:2,1:2] = NA
elev_coarse = getCoarseData(elevation, n:1, 1:n, FALSE, verbose=TRUE)
elev_ngb = getNeighbors(elev_coarse$data, verbose=TRUE)
getSimilarityMatrix(elevation, elev_ngb, TRUE)

}
