% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPath.R
\name{plotPath}
\alias{plotPath}
\title{Plot a path}
\usage{
plotPath(dataset, plot_data = FALSE, lat_bounds = NULL,
  lon_bounds = NULL, colored_line = "value",
  fixed_aspect_ratio = TRUE, background = NA,
  show_line_scale = FALSE, show_fill_scale = FALSE)
}
\arguments{
\item{dataset}{List with data, path, lat, lon, etc.}

\item{plot_data}{Plot the data?}

\item{lat_bounds}{Latitude bounds (for zooming in)}

\item{lon_bounds}{Longitude bounds (for zooming in)}

\item{colored_line}{Color the path: NA=>no, "group"=> by group ID, "value"=> by group value}

\item{fixed_aspect_ratio}{Use a 1:1 aspect ratio in the plot?}

\item{background}{Background color}

\item{show_line_scale}{Display colorbar for line?}

\item{show_fill_scale}{Display colorbar for fill?}
}
\value{
ggplot of path
}
\description{
Plot a path
}
