% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing function to add NAs in regions where values exceed a threshold}
\usage{
preprocess(data, thresh, verbose)
}
\arguments{
\item{data}{Raw data with no NAs}

\item{thresh}{Upper bound for valid values}

\item{verbose}{Print status to console?}
}
\value{
Data with NAs in appropriate regions
}
\description{
Preprocessing function to add NAs in regions where values exceed a threshold
}
