% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocessBathymetry}
\alias{preprocessBathymetry}
\title{Special preprocessing algorithm for bathymetric data}
\usage{
preprocessBathymetry(data, neritic, verbose)
}
\arguments{
\item{data}{Raw data with no NAs}

\item{neritic}{Depth (negative value) defining end of coastal region}
}
\value{
Scaled data (0-2) with NAs in appropriate regions, values
        between 0 and 1 in coastal regions, and between 1 and 2 in oceanic regions
}
\description{
Special preprocessing algorithm for bathymetric data
}
