% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sfc}
\alias{sfc}
\title{Fit a context-dependent surface filling curve to a dataset}
\usage{
sfc(data = preprocessBathymetry(etopo30, neritic = -1000),
  lat = seq(89.75, -89.75, by = -0.5), lon = seq(-179.75, 179.75, by =
  0.5), verbose = TRUE)
}
\arguments{
\item{data}{Dataset in matrix format}

\item{lat}{Vector of latitudes}

\item{lon}{Vector of longitudes}

\item{verbose}{Print status to console?}
}
\value{
Surface filling curve object (list)
}
\description{
Fit a context-dependent surface filling curve to a dataset
}
\examples{
sfc(data = preprocessBathymetry(etopo30, neritic = -1000),
    lat = seq(89.75, -89.75, by=-0.5),
    lon = seq(-179.75, 179.75, by=0.5),
    verbose = TRUE)

}
