% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPath.R
\name{plotPath}
\alias{plotPath}
\title{Plot a path}
\usage{
plotPath(dataset, plot_data = FALSE, lat_bounds = NULL,
  lon_bounds = NULL, colored_line = TRUE, fixed_aspect_ratio = TRUE,
  background = "black")
}
\arguments{
\item{dataset}{List with data, path, lat, lon, etc.}

\item{plot_data}{Plot the data?}

\item{lat_bounds}{Latitude bounds (for zooming in)}

\item{lon_bounds}{Longitude bounds (for zooming in)}

\item{colored_line}{Color the path using a rainbow palette?}

\item{fixed_aspect_ratio}{Use a 1:1 aspect ratio in the plot?}

\item{background}{Background color}
}
\value{
ggplot of path
}
\description{
Plot a path
}
